+++ src/multifloats-mpi/multifloats_mpi_f.f90
@@ -27,12 +27,6 @@ module multifloats_mpi_f
    public :: MPI_WW_SUM, MPI_WW_AMX, MPI_WW_AMN
    public :: multifloats_mpi_init
 
-   ! Deprecated pre-v0.14 names (arithmetic-based DD/ZZ); use the
-   ! family-prefixed MPI_MM_* / MPI_WW_* above. Bound to the C alias
-   ! symbols of the same handles. Remove after v0.14.
-   public :: MPI_DD_SUM, MPI_DD_AMX, MPI_DD_AMN
-   public :: MPI_ZZ_SUM, MPI_ZZ_AMX, MPI_ZZ_AMN
-
    integer(c_int), bind(c, name='mf_mpi_float64x2_f'),  protected :: MPI_FLOAT64X2
    integer(c_int), bind(c, name='mf_mpi_complex64x2_f'),protected :: MPI_COMPLEX64X2
    integer(c_int), bind(c, name='mf_mpi_mm_sum_f'),     protected :: MPI_MM_SUM
@@ -41,13 +35,6 @@ module multifloats_mpi_f
    integer(c_int), bind(c, name='mf_mpi_ww_sum_f'),     protected :: MPI_WW_SUM
    integer(c_int), bind(c, name='mf_mpi_ww_amx_f'),     protected :: MPI_WW_AMX
    integer(c_int), bind(c, name='mf_mpi_ww_amn_f'),     protected :: MPI_WW_AMN
-
-   integer(c_int), bind(c, name='mf_mpi_dd_sum_f'),     protected :: MPI_DD_SUM
-   integer(c_int), bind(c, name='mf_mpi_dd_amx_f'),     protected :: MPI_DD_AMX
-   integer(c_int), bind(c, name='mf_mpi_dd_amn_f'),     protected :: MPI_DD_AMN
-   integer(c_int), bind(c, name='mf_mpi_zz_sum_f'),     protected :: MPI_ZZ_SUM
-   integer(c_int), bind(c, name='mf_mpi_zz_amx_f'),     protected :: MPI_ZZ_AMX
-   integer(c_int), bind(c, name='mf_mpi_zz_amn_f'),     protected :: MPI_ZZ_AMN
 
    interface
       subroutine multifloats_mpi_init() bind(c, name='multifloats_mpi_init')